Move and rename createComponent.tsx.
